test(agent): move spec-sharing tests to a new part file
Relocate the spec-view sharing and dedup tests out of part 01 so the
builder test suite stays within its per-file size budget. The tests
themselves are unchanged.

Auto-committed-on: macbook
